backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
asyncpg
pydantic
httpx
python-dotenv
